pyproject.toml
[tool.poetry]
name = "eth-jit-exitter"
version = "0.1.0"
description = ""
authors = ["concentricspheres <107502292+concentricspheres@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "eth_jit_exitter"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
flask = "^2.2.3"
googleapis-common-protos = "^1.59.0"
grpcio = "^1.53.0"
pyyaml = "^6.0"
requests = "^2.28.2"
eth2spec = "^1.1.10"
setuptools = "^67.6.1"
aiohttp = "^3.8.4"
pyinstaller = "^5.10.1"
google-api-core = "^2.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
